no more hashes

+++ OpenDreamClient/Rendering/DreamViewOverlay.cs
@@ -562,7 +562,7 @@ internal sealed class DreamViewOverlay : Overlay {
         // Gather up all the data the view algorithm needs
         while (tileRefs.MoveNext(out var tileRef)) {
             var delta = tileRef.GridIndices - eyeTile.GridIndices;
-            var appearance = _appearanceSystem.GetTurfIcon(tileRef.Tile.TypeId).Appearance;
+            var appearance = _appearanceSystem.GetTurfIcon((uint)tileRef.Tile.TypeId).Appearance;
             if (appearance == null)
                 continue;
 
@@ -628,7 +628,7 @@ internal sealed class DreamViewOverlay : Overlay {
             tValue = 0;
             //pass the turf coords for client.images lookup
             Vector3 turfCoords = new Vector3(tileRef.X, tileRef.Y, (int) worldPos.MapId);
-            ProcessIconComponents(_appearanceSystem.GetTurfIcon(tileRef.Tile.TypeId), worldPos.Position - Vector2.One, EntityUid.Invalid, false, ref tValue, _spriteContainer, turfCoords: turfCoords);
+            ProcessIconComponents(_appearanceSystem.GetTurfIcon((uint)tileRef.Tile.TypeId), worldPos.Position - Vector2.One, EntityUid.Invalid, false, ref tValue, _spriteContainer, turfCoords: turfCoords);
         }
 
         // Visible entities
